statisticalFactsheets: added common: getLabel, levels
