+++ webapp/app/statisticalFactsheets/common/levels.js
@@ -1,0 +1,4 @@
+export const levels = {
+  global: 'WO',
+  region: ['AF', 'AS', 'EU', 'NA', 'OC', 'SA'],
+}
